[WIP][Prototype] Add GRPC streaming support to transport layer
Summary: This PR represents a first pass at what a streaming API would
look like in yarpc.  I'm modeling the API mostly off of the GRPC stream
APIs.

This has a few sections:

- Adding "Stream" as an option in the API surface (I've marked any
breaking changes we'll need to work around).
- Updating the Dispatcher to support streaming.
- Adding a raw encoding Stream API surface.
- Adding GRPC support to expose the streaming APIs.
- Adding a "streaming" example client/server that can be used to test
  the streaming APIs.

Since this PR does have breaking API changes in a few areas (*cough*
ClientConfig *cough*) it is not landable yet, but there's a lot of
feedback that can be given.

Test Plan: I've been using the "streaming" example to ensure that the
client and server can interact with each other, but, as we refine this
we'll need stricter testing.